requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.3
python-multipart==0.0.6
tiktoken==0.5.2
